access control moved to its own package. Tests stay in resolver to avoid invalide import cycle
