+++ pkg/accesscontrol/access_control.go
@@ -1,4 +1,4 @@
-package resolvers
+package accesscontrol
 
 import (
 	"fmt"
